algunas modificaciones para producto y mesero

+++ src/main/java/com/example/proydbp/mesero/dto/MeseroResponseDto.java
@@ -31,5 +31,5 @@ public class MeseroResponseDto {
 
     private List<ReviewMeseroResponseDto> reviewMeseros;
 
-    private Double ratingScore;
+    private Float ratingScore;
 }
